write.qvalue {qvalue}	R Documentation

Write results to file

Description

Write the results of the q-value object to a file.

Usage


write.qvalue(x, file = NULL, sep = " ", eol = "\n", na = "NA",
  row.names = FALSE, col.names = TRUE)

Arguments

x
A q-value object.

file
Output filename (optional).

sep
Separation between columns.

eol
Character to print at the end of each line.

na
String to use when there are missing values.

row.names
logical. Specify whether row names are to be printed.

col.names
logical. Specify whether column names are to be printed.

Details

The output file includes: (i) p-values, (ii) q-values (iii) local FDR values, and (iv) the estimate of pi_0, one per line. If an FDR significance level was specified in the call to qvalue, the significance level is printed and an indicator of significance is included.

Value

Nothing of interest.

Author(s)

John D. Storey, Andrew J. Bass

See Also

qvalue, plot.qvalue, summary.qvalue

Examples


# import data
data(hedenfalk)
p <- hedenfalk$p

# write q-value object
qobj <- qvalue(p)
write.qvalue(qobj, file="myresults.txt")

[Package qvalue version 2.14.1 Index]